+++ src/gql/ElementQueryConditionBuilder.php
@@ -486,7 +486,7 @@ class ElementQueryConditionBuilder extends Component
                         }
 
                         // For relational fields, prepare the arguments.
-                        if ($craftContentField instanceof BaseRelationField) {
+                        if ($craftContentField instanceof EagerLoadingFieldInterface) {
                             $arguments = $this->_argumentManager->prepareArguments($arguments);
                         }
                     }